move stock and search bar to wl folder
