+++ front-end/src/scenes/asset_screener/AssetScreener.tsx
@@ -1,10 +1,13 @@
+import {Box} from '@mui/material'
 import React from 'react'
 
 const AssetScreener = () => {
     return (
-        <div>
-            <h1>Asset Screener</h1>
-        </div>
+        <>
+            <Box>
+                <h1>Asset Screener</h1>
+            </Box>
+        </>
     )
 }
 export default AssetScreener
