+++ front-end/src/scenes/dashboard/Dashboard.tsx
@@ -3,7 +3,7 @@ import {Box} from '@mui/system'
 import {useEffect, useState} from 'react'
 import Cookies from 'universal-cookie'
 import Watchlist from '../watchlist/Watchlist'
-import Searchbar from '../../components/UI/Searchbar'
+import Searchbar from '../watchlist/stock/Searchbar'
 import {getUserWL} from '../../config/WebcallAPI'
 import CreateWLButton from '../watchlist/WL_Action_Buttons/Create_Button'
 import AddTickersButton from '../watchlist/WL_Action_Buttons/Add_Tickers_Button'
