+++ front-end/src/scenes/watchlist/WL_Action_Buttons/Delete_Tickers_Button.tsx
@@ -2,7 +2,7 @@ import {useState} from 'react'
 import {delTickersFromWL} from '../../../config/WebcallAPI'
 import {Box, Button, Modal, Typography, colors} from '@mui/material'
 import modalStyle from '../WatchlistStyles'
-import Searchbar from '../../../components/UI/Searchbar'
+import Searchbar from '../stock/Searchbar'
 
 type DeleteTickersButtonProps = {
     user_id: string
