+++ front-end/src/scenes/watchlist/Watchlist.tsx
@@ -9,12 +9,12 @@ import TableContainer from '@mui/material/TableContainer'
 import TableHead from '@mui/material/TableHead'
 import TableRow from '@mui/material/TableRow'
 import {useContext, useEffect, useState} from 'react'
-import Stock from '../../components/stock/Stock'
+import Stock from './stock/Stock'
 import {ColorModeContext, tokens} from '../../theme'
 import DisplayGroup from './DisplayGroup'
 import Cookies from 'universal-cookie'
 import modalStyle from './WatchlistStyles'
-import Searchbar from '../../components/UI/Searchbar'
+import Searchbar from './stock/Searchbar'
 import {addTickersToWL, createWL, delTickersFromWL, deleteWL, getWLAssets, renameWL} from '../../config/WebcallAPI'
 
 // TODO:
